Call refactor command with default language extensions.
Out of the box the 'refactor' command doesn't work with syntax altering
extensions like LambdaCase and RankNTypes. To enable the right behaviour
'refactor' expects the extensions as -X... arguments or directly in the
module as LANGUAGE pragmas.

This change lets the plugin work with default extensions defined with hpack
(which is often used in conjunction with stack) by parsing the
extensions from hpacks package.yaml and providing them as arguments to
the 'refactor' command.

TODO: The implementation is missing the ability to update the extensions, when
the package.yaml is changed.